Add inspection-batch-with-output-gathering example
This workflow submits a Job which creates `n` pod replicas and runs them in a sequence.
Additionally, it mounts an artifact repository to store the resulting artifacts.

NOTE: The artifact repository has to be configured via ConfigMap, see
https://github.com/argoproj/argo/blob/master/ARTIFACT_REPO.md
or use the Argo ansible role: https://github.com/CermakM/ansible-role-argo-workflows

successCondition: `n` pods have finished successfully and are present in the cluster

The workflow simulates the scenario when we want to create a batch of inspections
and store the outputs to the artifact store.

Signed-off-by: Marek Cermak <macermak@redhat.com>